move exit to an internal directory so it cant be used outside of logging
