+++ pkg/logging/internal/exit/exit.go
