+++ pkg/logging/internal/exit/exit_test.go
